fixed per Janoses comments
