+++ pkg/p2p/p2p.go
@@ -37,7 +37,7 @@ type ProtocolSpec struct {
 	Name        string
 	Version     string
 	StreamSpecs []StreamSpec
-	Init        func(peer Peer) error
+	Init        func(ctx context.Context, peer Peer) error
 }
 
 type StreamSpec struct {
